+++ packages/webextension/app/scripts/InstalledApp/InstalledApp.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { InstalledTextlintList } from "./component/InstalledTextlintList";
+
+export const InstalledApp = () => {
+    return (
+        <div className={"InstalledApp"}>
+            <InstalledTextlintList url={location.href} />
+        </div>
+    );
+};